New URL Service for Radio and Watch.cbc.ca
CBC Radio moved to a new URL — and presumably to a new API — so this
required a rewrite.

Many video categories also moved to a new video player; wrote a new URL
service for that.

Existing URL service is not changed, other than a few comments to help
with dev/debugging